Combine device and distributed utils to avoid circular imports
