+++ hydragnn/models/__init__.py
@@ -3,4 +3,4 @@ from .GATStack import GATStack
 from .GINStack import GINStack
 from .PNAStack import PNAStack
 from .GINStack import GINStack
-from .create import create, get_device
+from .create import create
