+++ hydragnn/utils/__init__.py
@@ -1,6 +1,9 @@
 from .print_utils import print_distributed, iterate_tqdm
-from .device import get_device_list, get_device
-from .distributed import get_comm_size_and_rank
+from .distributed import (
+    get_comm_size_and_rank,
+    get_device_list,
+    get_device,
+)
 from .time_utils import Timer, print_timers
 from .config_utils import (
     update_config_NN_outputs,
